% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asmr}
\alias{asmr}
\title{Age specific mortality rates}
\format{
A data frame
}
\source{
UNWPP, wrangled into a csv file by the VIMC team at
Imperial, distributed through Montagu.
}
\usage{
asmr
}
\description{
A list of all age smoothed mortality rates currently available from the
UNWPP, who released the data with a free license to use and distribute as
desired, with modifications. Year indicated is from the start of a 5 year
period of estimation, (i.e. values associated with 1950 indicates values
averaged over 1950.5 until 1955.5). Ages ranges are inclusive of the age
indicated in age_to. The death rates are from the mx values of the UNWPP
lifetables, and further details on how they were computed can be found at
\url{https://www.un.org/development/desa/pd/data/model-life-tables}.
}
\keyword{datasets}
