% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nmr_asmr.R
\name{asmr_to_nmr}
\alias{asmr_to_nmr}
\title{Function to convert ASMR to NMR format}
\usage{
asmr_to_nmr(asmr = peeps::asmr)
}
\arguments{
\item{asmr}{Data frame of age standardized mortality rates.}
}
\value{
Dataframe
}
\description{
Break the 5 year averages into single year values.
}
