% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbr}
\alias{cbr}
\title{Crude birth rates}
\format{
A data frame with 17100 rows and 8 variables:
\describe{
\item{country_code}{Iso3c codes}
\item{country}{Country name}
\item{age_from}{Starting age of the age band, inclusive.}
\item{age_to}{Ending age of the age band, inclusive. Is 120 for all entries.}
\item{year}{Year applicable, ranges from 1950-2099}
\item{gender}{"both" genders only.}
\item{value}{Number of births over a given period divided by the person-
years lived by the population over that period. It is expressed as average
annual number of births per 1,000 population.}
\item{income_group}{World Bank country income group.}
}
}
\source{
UNWPP
}
\usage{
cbr
}
\description{
Crude birth rates
}
\keyword{datasets}
