% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data
Updated by Ines on 14 feb.}
\usage{
download_data(indicator = "deathrates")
}
\arguments{
\item{indicator}{Deathrates or population.}
}
\value{
Either list of dataframes containing deathrates, or dataframe of
population by year.
}
\description{
Download data
Updated by Ines on 14 feb.
}
