% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{estimate_mortality_rates}
\alias{estimate_mortality_rates}
\title{Estimate mortality rates}
\usage{
estimate_mortality_rates(target_age_distribution, starting_mortality_rates)
}
\arguments{
\item{target_age_distribution}{Target age distribution. The proportion of the
population in each age band. This should sum to 1.}

\item{starting_mortality_rates}{A vector of mortality rates to initialise the optimisation.
This is recommended to be the variable qx from the UN WPP life tables}
}
\value{
The sum of squared  differences between the target and current age distributions
}
\description{
Optimises the mortality rates to reproduce a target (equilibrium) age distribution
}
