% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifetable.R
\name{extract_deathrates}
\alias{extract_deathrates}
\title{Get deathrates if input a life table returned from the \code{\link[=get_lifetable]{get_lifetable()}} function}
\usage{
extract_deathrates(data)
}
\arguments{
\item{data}{data frame of life table data returned from \code{\link[=get_lifetable]{get_lifetable()}}}
}
\value{
Data frame of death rates
}
\description{
Get deathrates if input a life table returned from the \code{\link[=get_lifetable]{get_lifetable()}} function
}
\examples{
extract_deathrates(DT)
}
\keyword{UN,}
\keyword{age}
\keyword{ax,}
\keyword{deathrates,}
\keyword{demography,}
\keyword{download,}
\keyword{dx,}
\keyword{gender,}
\keyword{groups}
\keyword{life}
\keyword{mx,}
\keyword{population,}
\keyword{table,}
