% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_mortality.R
\name{get_child_mortality}
\alias{get_child_mortality}
\title{Get child mortality rates}
\usage{
get_child_mortality(
  country = "all",
  gender = "all",
  years.start = 1960,
  years.end = NA,
  indicator = "all"
)
}
\arguments{
\item{country}{all, iso3c code, or country name in English. If not available or
misspelled, will throw error. To do: add a list of all countries/regions that we
have data for.}

\item{gender}{both (only total), female, male, all (female, male, and total).
Data by gender is only available for infant mortality and under 5 mortality.}

\item{years.start}{Mortality data is available between 1960-2019. This specifies
the first year for which values are returned, default is 1960.}

\item{years.end}{Specified the last year for which values are returned. By default
returns up to 2019.}

\item{indicator}{Indicators for the function to return. Either returns all,
infant mortality ("infant"), neonatal mortality("neonatal"), or under 5
mortality ("under 5").}
}
\value{
List of mortality rates.
}
\description{
This function computes and returns the child mortality rates developed by the
UN Inter-Agency Group for Child Mortality Estimation, which includes UNICEF,
WHO, the World Bank, and the UN DESA Population Division. This data is retrieved
from the World Bank website, and terms of use are found here: https://www.worldbank.org/en/about/legal/terms-of-use-for-datasets.
}
\details{
Neonatal mortality: the probability of dying between the first 28 days of life, expressed per 1000 live births.
Infant mortality: the probability of dying between birth and exact age 5, expressed per 1000 live births.
Under 5 mortality rate: the probability of dying between birth and exact age 5, expressed per 1000 live births.

Future additions:
Memory use improvements. Check of years and ages inputs. Uses at least an extra 2.6 MB.
Spell checks. Url check - make sure its still there.
}
\examples{
get_child_mortality(country="Norway")
}
\keyword{5}
\keyword{UN,}
\keyword{child,}
\keyword{deathrates,}
\keyword{demography,}
\keyword{download,}
\keyword{infant,}
\keyword{mortality,}
\keyword{neonatal,}
\keyword{population,}
\keyword{under}
