% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imr}
\alias{imr}
\title{Infant mortality rates}
\format{
A data frame with 17100 rows and 8 variables:
\describe{
\item{country_code}{Iso3c codes}
\item{country}{Country name}
\item{age_from}{Starting age of the age band, inclusive.}
\item{age_to}{Ending age of the age band, inclusive. Is 1 for all infants.}
\item{year}{Year applicable, ranges from 1950-2100}
\item{gender}{"both" genders only.}
\item{value}{Probability of dying between birth and exact age 1 year. It is
expressed as average annual deaths per 1,000 births.}
\item{income_group}{World Bank country income group.}
}
}
\source{
UNWPP.
}
\usage{
imr
}
\description{
Infant mortality values provided by the UNWPP.
}
\keyword{datasets}
