% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_deathrates.R
\name{load_deathrates}
\alias{load_deathrates}
\title{Load deathrates}
\usage{
load_deathrates(iso3c = NULL, country = NULL, year = NULL, format = "asmr")
}
\arguments{
\item{iso3c}{Country code to subset deathrates by.}

\item{country}{Country name to subset deathrates by.}

\item{year}{Year to subset deathrates by. Can be single value or vector.
If format = 'asmr', year data is given in 5 year chunks. If format = 'nmr',
year is available by single year. There is no difference in the distribution
of values between the two formats.}

\item{format}{Format to return deathrates by ('asmr' = by years divisible by
5, or 'nmr' = single year values). See 'peeps::merge_nmr_asmr' documentation
for details.}
}
\value{
Data.frame of deathrates for country.
}
\description{
Taking the input of a country or iso3c code, load the deathrates in
the format required by \href{https://github.com/mrc-ide/malariasimulation/}{malariasimulation}.
}
