% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nmr_asmr.R
\name{merge_nmr_asmr}
\alias{merge_nmr_asmr}
\title{Merge neonatal deathrates with population deathrates.}
\usage{
merge_nmr_asmr(format = "asmr")
}
\arguments{
\item{format}{Format to convert dataset to. Can be either asmr or nmr format.}
}
\value{
Dataframe of death rates from the ages of 0-120, from 1950-2100.
}
\description{
While the neonatal mortality rates (NMR) are estimates by year, the
age specific mortality rates (ASMR) are given per group of 5 years. The user
can combine these functions a few different ways:
}
\details{
\enumerate{
\item Take the 5 year averages of the neonatal rates, so transform them to be of
the same format as the death rates, then combine (format = "asmr", uses the
nmr_to_asmr() function).
\item Expand the death rates data -> simply copy the values from 1950 onto years
1951, 1952, 1953, and 1954, etc. then combine with neonatal data (format =
"nmr", uses the asmr_to_nmr() function).
}

As the NMR rates are not available segregated by gender, only rates for the
full population are returned.

This output can be subset by country and year to be directly fed into
malariasimulation.
}
