% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nga_example}
\alias{nga_example}
\title{Example UN WPP mortality and age-structure data}
\format{
An object of class \code{data.frame} with 101 rows and 6 columns.
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Mortality/}
}
\usage{
nga_example
}
\description{
A data frame with 101 rows and 6 columns:
\describe{
\item{region}{Country name}
\item{iso3x}{3 letter ISO country codes}
\item{age_lower}{Lower end of age group (years)}
\item{age_upper}{Upper end of age group (years)}
\item{qx}{Probability of dying in age group}
\item{p}{Proportion of population in age group}
}
}
\keyword{datasets}
