% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr}
\alias{nmr}
\title{Neonatal mortality rates}
\format{
A data frame
}
\source{
Child mortality project and UNWPP.
}
\usage{
nmr
}
\description{
The Imperial team calculated the NMR by taking the NMR/IMR ratio from the
child mortality dataset (see \url{childmortality.org} for more details) and
applying that transformation to the IMR values calculated in the life tables
of the UNWPP. This in order to keep data consistent and at the UNWPP scale,
while incorporating the higher accuracy and granularity of the child
mortality project estimates.
}
\keyword{datasets}
