% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nmr_asmr.R
\name{nmr_to_asmr}
\alias{nmr_to_asmr}
\title{Function to convert NMR to ASMR format}
\usage{
nmr_to_asmr(nmr = peeps::nmr)
}
\arguments{
\item{nmr}{Dataframe of neonatal mortality rates.}
}
\value{
Dataframe
}
\description{
What this means is that it will the 5 year average
}
