% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{optim_objective}
\alias{optim_objective}
\title{optimisation objective function}
\usage{
optim_objective(mortality_rates, target_age_distribution)
}
\arguments{
\item{mortality_rates}{Mortality rates}

\item{target_age_distribution}{Target age distribution}
}
\value{
The sum of squared  differences between the target and current age distributions
}
\description{
optimisation objective function
}
