% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{Population basic dataset from 1950-2099.}
\format{
A data frame with 17214 rows and 8 variables:
\describe{
\item{country_code}{Iso3c codes}
\item{country}{Country name}
\item{age_from}{Starting age of the age band, inclusive.}
\item{age_to}{Ending age of the age band, inclusive. Is 120 for all entries.}
\item{year}{Year applicable, ranges from 1950-2099}
\item{gender}{"both" genders only.}
\item{value}{Number of people within that category.}
\item{income_group}{World Bank country income group.}
}
}
\source{
UNWPP
}
\usage{
pop
}
\description{
Population basic dataset from 1950-2099.
}
\keyword{datasets}
