% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_detail}
\alias{pop_detail}
\title{Population detailed dataset - in 5 year age groups and over 5 year time spans.}
\format{
A data frame with 249318 rows and 8 variables:
\describe{
\item{country_code}{Iso3c codes}
\item{country}{Country name}
\item{age_from}{Starting age of the age band, inclusive. Is 0 for <1 year. Max
years is 120.}
\item{age_to}{Ending age of the age band, inclusive. Is 0 for <1 year. Max
years is 120.}
\item{year}{Starting year of 5 year group, i.e. 1960 for 1960-64, etc. From
1950-2100.}
\item{gender}{"both", "female", or "male"}
\item{value}{Number of people within that category.}
\item{income_group}{World Bank country income group.}
}
}
\source{
UNWPP
}
\usage{
pop_detail
}
\description{
Population detailed dataset - in 5 year age groups and over 5 year time spans.
}
\keyword{datasets}
