% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{rescale_prob}
\alias{rescale_prob}
\title{Convert a probability between time intervals}
\usage{
rescale_prob(p, interval_in, interval_out)
}
\arguments{
\item{p}{Probability}

\item{interval_in}{Interval in}

\item{interval_out}{Interval out (in same time units as interval in)}
}
\value{
Rescaled probability
}
\description{
Convert a probability between time intervals
}
