% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asmr}
\alias{asmr}
\title{Age specific mortality rates}
\format{
A data frame with 225720 rows and 7 variables:
\describe{
\item{country_code}{Iso3c codes}
\item{country}{Country name}
\item{age_from}{Starting age of the age band, inclusive.}
\item{age_to}{Ending age of the age band, inclusive. Is 0 for <1 year. Max
years is 120.}
\item{year}{Starting year of 5 year group, i.e. 1960 for 1960-64, etc. From
1950-2100.}
\item{gender}{"both", "female", or "male"}
\item{value}{Age specific mortality rate, or the central rate of
mortality, mx.}
}
}
\source{
UNWPP, wrangled into a csv file by the VIMC team at
Imperial, distributed through Montagu.
}
\usage{
asmr
}
\description{
A list of all age smoothed mortality rates currently available from the
UNWPP, who released the data with a free license to use and distribute as
desired, with modifications. Year indicated is from the start of a 5 year
period of estimation, (i.e. values associated with 1950 indicates values
averaged over 1950.5 until 1955.5). Ages ranges are inclusive of the age
indicated in age_to. The death rates are from the mx values of the UNWPP
lifetables, and further details on how they were computed can be found at
\url{https://www.un.org/development/desa/pd/data/model-life-tables}.
}
\keyword{datasets}
