% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_deathrates.R
\name{load_deathrates}
\alias{load_deathrates}
\title{Load deathrates}
\usage{
load_deathrates(iso3c = NULL, country = NULL, year = NULL)
}
\arguments{
\item{iso3c}{Country code to subset deathrates by.}

\item{country}{Country name to subset deathrates by.}

\item{year}{Year to subset deathrates by. Can be single value or vector.}
}
\value{
Data.frame of deathrates for country.
}
\description{
Taking the input of a country or iso3c code, load the deathrates in
the format required by \href{https://github.com/mrc-ide/malariasimulation/}{malariasimulation}.
}
