% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nmr_asmr.R
\name{merge_nmr_asmr}
\alias{merge_nmr_asmr}
\title{Merge neonatal deathrates with population deathrates.}
\usage{
merge_nmr_asmr(format = "asmr")
}
\arguments{
\item{format}{Format to convert dataset to. Can be either asmr or nmr format.}
}
\value{
Vector of deathrates by year.
}
\description{
While the neonatal mortality rates are estimates by year, the
age specific mortality rates are given per group of 5 years. The user
can combine these functions a few different ways:
}
\details{
\enumerate{
\item Take the 5 year averages of the neonatal rates, so transform them to be of
the same format as the death rates, then combine.
\item Expand the death rates data -> simply copy the values from 1950 onto years
1951, 1952, 1953, and 1954. Repeat onwards. Combine with neonatal data.
\item Expanding the death rate data, including an element of stochasticity, to
be discussed.
}
}
