% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nmr_asmr.R
\name{nmr_to_asmr}
\alias{nmr_to_asmr}
\title{Function to convert NMR to ASMR format}
\usage{
nmr_to_asmr(nmr = peeps::nmr)
}
\arguments{
\item{nmr}{Dataframe of neonatal mortality rates.}
}
\value{
Dataframe of 5 year averages of neonatal mortality rates
}
\description{
This function returns the average of the NMR for every five year band, i.e.
for 1950-1954, 1955-1959, etc., so that it can be combined with the ASMR.
}
